+++ Source/Modules/Catalog/Security/Gapotchenko.FX.Security.Cryptography/DESUnapprovedImpl.cs
@@ -6,11 +6,11 @@
 // File introduced by: Oleksiy Gapotchenko
 // Year of introduction: 2026
 
-#pragma warning disable CA5351 // Do Not Use Broken Cryptographic Algorithms
-
 using Gapotchenko.FX.Security.Cryptography.Kits;
 using Gapotchenko.FX.Security.Cryptography.Properties;
 using System.Runtime.CompilerServices;
+
+#pragma warning disable CA5351 // Do Not Use Broken Cryptographic Algorithms
 
 namespace Gapotchenko.FX.Security.Cryptography;