+++ Documentation/Changelog.md
@@ -6,6 +6,7 @@
 
 Release date: not released yet
 
+- Added DES and MD5 algorithm implementations that can be used for compatibility with legacy systems and data formats
 - Polyfills:
     - Added polyfill for `System.Security.Cryptography.CryptographicOperations` class
     - Added polyfills for `Fill`, `GetBytes`, and `GetInt32` methods of `System.Security.Cryptography.RandomNumberGenerator` class